Extract shared fs code into shared.ts, simplify getDataDir on web
